day09/lab/requirements.txt
# Day 09 Lab — Multi-Agent Orchestration
# Install: pip install -r requirements.txt

# ─── Core ───────────────────────────────
python-dotenv>=1.0.0

# ─── Vector Store ───────────────────────
chromadb>=0.4.22
qdrant-client>=1.8.0

# ─── Embeddings ─────────────────────────
sentence-transformers>=2.6.0

# ─── LLM ────────────────────────────────
openai>=1.12.0
google-generativeai>=0.5.0
langchain-google-genai>=1.0.0

# ─── Orchestration ──────────────────────
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0

# ─── Data ───────────────────────────────
pyyaml>=6.0.1

# ─── MCP (optional — Sprint 3 Bonus) ───
mcp>=1.0.0
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.27.0

# ─── Dev/Test ───────────────────────────
pytest>=8.0.0
